requirements.txt
ib_insync
slackclient==2.9.4
